move translatable base classes
